Crate openrailwaymap_exporter

OpenRailwayMap Exporter

A command-line tool built with Rust that fetches railway data from OpenRailwayMap using the Overpass API and exports it into a custom format.

Features

Downloads railway data using Overpass API

Retrieves railway track data including length, GPS path, IDs, connected elements, and switches

Accepts bounding box parameter to define the area for which data should be downloaded

Data will be downloaded from overpass-turbo.

Usage

Run the command-line tool using the following command structure:

cargo run -- --bbox "latitude_min,longitude_min,latitude_max,longitude_max"

For example, to download railway data within a bounding box around Frankfurt, use:

cargo run -- --bbox "49.9,8.4,50.2,8.8"

or use the following:

cargo run -- --area "Frankfurt am Main"

To save the elements in a json file, use the following:

cargo run -- --area "Frankfurt am Main" -j -o output.json

To save the graph as svg format use:

cargo run -- --area "Frankfurt am Main" -d -o output.dot

then you can convert the graphviz format into a svg image using the following command:

cargo run -- --area "Frankfurt am Main" --svg -o output.svg

Web App

Compile the code to WASM:

wasm-pack build --target web

Run the Web version in your browser

python3 -m http.server

Open your browser on Localhost

Running Tests

To run tests, use the following command:

cargo test

Running Benchmarks

To run benchmarks, use the following command:

cargo bench

Profiling with Flamegraph

To create a flamegraph for a specific example, first, make sure you have cargo-flamegraph installed:

cargo install flamegraph

Then, run the following command:

cargo flamegraph --dev --example vilbel

This will generate a flamegraph.svg for the vilbel example.

Contributing

Contributions are welcome! Please feel free to submit issues or pull requests.

License

This project is licensed under the MIT License. See the LICENSE file for details.

Modules

railway_graph

A module for working with railway graphs.

Structs

BasicOpenRailwayMapApiClient

A basic client for the OpenRailwayMap API.

Coordinate

A Coordinate represents a geographic coordinate with latitude and longitude.

RailwayEdge

Represents a railway edge with a unique ID and a length.

RailwayElement

Represents an element of a railway network.

RailwayGraph

RailwayGraph represents a graph structure for railway networks.

RailwayNode

Represents a railway node with a unique ID and geographic coordinates (latitude and longitude).

Enums

ElementType

Represents the type of a RailwayElement.

Traits

OpenRailwayMapApiClient

A trait for implementing an OpenRailwayMap API client.

Functions

count_node_elements

Counts the number of Node elements in a given vector of RailwayElements.

count_way_elements

Counts the number of Way elements in a vector of RailwayElements.

create_nodes

Create a vector of RailwayNodes from the provided RailwayElements.

find_next_existing_node

Find the next existing node ID and its index in the node_indices HashMap after the given start ID.

generate_dot_string

Generates a DOT string representation of a given RailwayGraph.

generate_svg_string

Generates an SVG string representation of a given RailwayGraph.