Struct openrailwaymap_exporter::ai::TrainAgentAI

pub struct TrainAgentAI {
    pub id: RailwayObjectId,
    pub railway_graph: Option<RailwayGraph>,
    pub current_node: Option<NodeId>,
    pub target_node: Option<NodeId>,
    pub agent_rl: TrainAgentRL,
    pub trainer: Arc<RwLock<AgentTrainer<TrainAgentState>>>,
}

A reinforcement learning agent that controls a train in the simulation.

Fields
id: RailwayObjectId

The id of the railway object

railway_graph: Option<RailwayGraph>

The railway graph representing the train network.

current_node: Option<NodeId>

The current node

target_node: Option<NodeId>

The target node

agent_rl: TrainAgentRL

The reinforcement learning agent responsible for controlling the train.

trainer: Arc<RwLock<AgentTrainer<TrainAgentState>>>

The trainer responsible for training the reinforcement learning agent.

Implementations

impl TrainAgentAI

pub fn new(railway_graph: RailwayGraph, initial_state: TrainAgentState) -> Self

Creates a new TrainAgentAI with the given railway graph and initial state.

Arguments

railway_graph - The railway graph representing the train network.

initial_state - The initial state of the train agent in the simulation.

Returns

A new TrainAgentAI instance.

pub fn train(&mut self, iterations: usize)

Trains the reinforcement learning agent for the specified number of iterations.

Arguments

iterations - The number of iterations to train the agent.

pub fn best_action(&self, state: &TrainAgentState) -> Option<RailMovableAction>

Returns the best action for the given state according to the trained reinforcement learning agent.

Arguments

state - The current state of the train agent in the simulation.

Returns

The best action for the given state or None if no action can be selected.

pub fn observe( &mut self, current_node: NodeId, target_node: Option<NodeId>, speed_mm_s: Option<i32>, delta_distance_mm: Option<i32> )

Updates the current edge, target node (optionally), and calculates the new state by updating its distance using the shortest path distance while keeping the current speed constant.

Arguments

current_edge - The current edge on which the train agent is.

target_node - The optional target node the train agent is heading to.

time_delta_ms - The time delta in milliseconds since the last update.