Struct openrailwaymap_exporter::ai::TrainAgentState

pub struct TrainAgentState {
    pub delta_distance_mm: i32,
    pub current_speed_mm_s: i32,
    pub max_speed_percentage: i32,
}

Represents the state of a train agent in the simulation.

Fields
delta_distance_mm: i32

The remaining distance in millimeters the train agent needs to travel.

current_speed_mm_s: i32

The current speed of the train agent in millimeters per second (mm/s).

max_speed_percentage: i32

The maximum speed percentage the train agent can reach (e.g., 100 for 100% of the maximum speed).