openrailwaymap_exporter

Module algorithms

A module containing algorithms for working with geographical data structures.

Traits

Distance

A trait that defines a method for calculating the distance between two points of the same type. Implementations are provided for Coord<f64> and Point<f64>.

Functions

closest_point_in_linestring

Returns the closest point of points in a given LineString to a given Coord.

is_middle_coord_between

Determines if the middle coordinate is between the start and end coordinates along both x and y axes.

points_in_front

Returns the points in front of the current location in a given LineString.