Function openrailwaymap_exporter::algorithms::is_middle_coord_between

pub fn is_middle_coord_between(
    start_coord: Coord<f64>,
    middle_coord: Coord<f64>,
    end_coord: Coord<f64>
) -> bool

Determines if the middle coordinate is between the start and end coordinates along both x and y axes.

This function assumes the three coordinates are collinear.

Arguments

start_coord - A Coord<f64> representing the start coordinate.

middle_coord - A Coord<f64> representing the middle coordinate.

end_coord - A Coord<f64> representing the end coordinate.

Returns

A bool indicating whether middle_coord is between start_coord and end_coord.