openrailwaymap_exporter::algorithms

Trait Distance

pub trait Distance {
    // Required method
    fn distance(&self, other: &Self) -> Length;
}

A trait that defines a method for calculating the distance between two points of the same type. Implementations are provided for Coord<f64> and Point<f64>.

Required Methods

fn distance(&self, other: &Self) -> Length

Calculate the distance between self and other and return the result as a Length in meters.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.