openrailwaymap_exporter::app3d

Function setup_app

pub fn setup_app(app: &mut App, app_resource: AppResource)

Configures a Bevy application with the necessary plugins, resources, and systems.

This function configures a Bevy application with the required plugins, resources, and systems based on the given AppResource. The caller is responsible for inserting any additional resources and running the application.

Arguments

app - A mutable reference to a bevy::prelude::App instance to configure.

app_resource - An AppResource to configure the application.