openrailwaymap_exporter::importer

Module overpass_importer

OverpassImporter is a struct that implements the RailwayGraphImporter trait for importing railway graph data from the Overpass API.

Structs

Coordinate

A Coordinate represents a geographic coordinate with latitude and longitude.

OverpassImporter

OverpassImporter is a struct that implements the RailwayGraphImporter trait for importing railway graph data from the Overpass API.

RailwayElement

Represents an element of a railway network.

Enums

ElementType

Represents the type of a RailwayElement.

Functions

calculate_geometry_length

Calculate the total length of a sequence of coordinates by summing the distance between consecutive coordinates.

count_node_elements

Counts the number of Node elements in a given vector of RailwayElements.

count_way_elements

Counts the number of Way elements in a vector of RailwayElements.

create_nodes

Create a vector of RailwayNodes from the provided RailwayElements.

create_nodes_from_way_elements_without_existing

Create railway nodes from two railway elements that share a node ID but do not have an existing common node. The nodes are created by finding the two coordinates in the geometries of the two railway elements that are closest to each other.

find_next_existing_node

Find the next existing node ID and its index in the node_indices HashMap after the given start ID.

from_railway_elements

Create a RailwayGraph from a vector of RailwayElements.