openrailwaymap_exporter

Module railway_algorithms

Module railway_algorithms provides algorithms for working with railway networks.

The module contains the PathFinding trait, which offers methods to calculate the shortest path distance, the shortest path as a list of node IDs, and the shortest path as a list of edge IDs for railway networks.

Traits

PathFinding

PathFinding trait provides pathfinding algorithms for railway networks.

RailwayEdgeAlgos

Algorithms for railway edges.

RailwayGraphAlgos

RailwayGraphAlgos trait provides algorithms for railway graphs.