Module openrailwaymap_exporter::railway_api_client

Railway API Client module for the OpenRailwayMap Exporter.

This module provides a trait and an implementation for fetching railway infrastructure data from an API. The RailwayApiClient trait provides a common asynchronous interface for fetching data by area name or bounding box.

Re-exports

pub use overpass_api_client::OverpassApiClient;

Modules

overpass_api_client

overpass api client

Traits

RailwayApiClient

A trait for implementing an Railway API client.