Module openrailwaymap_exporter::railway_api_client::overpass_api_client

overpass api client

Structs

Coordinate

A Coordinate represents a geographic coordinate with latitude and longitude.

OverpassApiClient

A basic client for the OpenRailwayMap API.

RailwayElement

Represents an element of a railway network.

Enums

ElementType

Represents the type of a RailwayElement.

Functions

count_node_elements

Counts the number of Node elements in a given vector of RailwayElements.

count_way_elements

Counts the number of Way elements in a vector of RailwayElements.