Struct openrailwaymap_exporter::railway_model::railway_graph::RailwayGraph

pub struct RailwayGraph {
    pub graph: Graph<RailwayNode, RailwayEdge, Undirected>,
    pub node_indices: HashMap<i64, NodeIndex>,
}

RailwayGraph represents a graph structure for railway networks.

The graph consists of nodes representing railway stations and junctions, and edges representing the railway segments between the nodes. It also stores the node indices as a HashMap for easy retrieval.

Fields
graph: Graph<RailwayNode, RailwayEdge, Undirected>

The internal graph used to represent the railway network.

The graph consists of RailwayNode instances as nodes and RailwayEdge instances as edges. It is an undirected graph.

node_indices: HashMap<i64, NodeIndex>

A HashMap that maps node IDs to their corresponding indices in the graph.

This HashMap allows for quick and easy retrieval of node indices based on their IDs.

Implementations

impl RailwayGraph

pub fn reachable_nodes(&self, start_node_id: i64) -> Vec<i64>

Find all reachable nodes from the given start node in the railway graph.

This function performs a breadth-first search from the given start node and returns a vector of reachable node IDs.

Arguments

start_node_id - The ID of the start node.

Returns

A Vec<i64> containing the IDs of all nodes reachable from the start node. If the start node ID is not found in the graph, an empty vector is returned.

pub fn get_next_node(&self, current: i64, target: i64) -> Option<i64>

Returns the next reachable node on the shortest path

impl RailwayGraph

pub fn get_edge_by_id(&self, id: i64) -> Option<RailwayEdge>

Retrieve an edge from the graph by its ID.

Arguments

id - The ID of the edge to be retrieved.

Returns

An Option<RailwayEdge> that contains the edge if found, or None if not found.

pub fn railway_edge( &self, start_node_id: i64, end_node_id: i64 ) -> Option<&RailwayEdge>

Retrieve the railway edge between two nodes.

Arguments

start_node_id - The ID of the starting node.

end_node_id - The ID of the ending node.

Returns

An Option<&RailwayEdge> that contains the railway edge connecting the two nodes if it exists, or None if no such edge exists.

pub fn bounding_box(&self) -> (Coord, Coord)

Calculate the bounding box of the graph.

The bounding box is represented as a tuple containing the minimum and maximum latitude and longitude values of the nodes in the graph.

Returns

A tuple containing two Coordinate structs representing the minimum and maximum coordinates of the bounding box of the graph.

pub fn total_length(&self) -> f64

Calculate the total length of the railway network.

The total length is the sum of the lengths of all edges in the graph.

Returns

A f64 value representing the total length of the railway network in meters.

pub fn nearest_node( &self, edge_id: i64, position_on_edge: f64, current_node_id: Option<NodeId> ) -> Option<NodeId>

Returns the nearest node to the given position on the specified edge.

Arguments

edge_id - The ID of the edge.

position_on_edge - The position on the edge, ranging from 0.0 to 1.0.

current_node_id - An optional NodeId of the current node to determine the start node.

Returns

An Option<NodeId> containing the ID of the nearest node if found, or None if the edge is not found.