openrailwaymap_exporter::railway_model

Type Alias RailwayEdge

pub type RailwayEdge = TransitEdge<f64>;

A railway edge.

Aliased Type

pub struct RailwayEdge {
    pub id: u64,
    pub source: u64,
    pub target: u64,
    pub length: f64,
    pub path: LineString,
}

Fields
id: u64

A unique identifier for the edge.

source: u64

The identifier of the node where the edge starts.

target: u64

The identifier of the node where the edge ends.

length: f64

The length of the edge.

path: LineString

The path of the edge, represented as a LineString.