openrailwaymap_exporter::railway_objects

Trait NextTarget

pub trait NextTarget: RailwayObject {
    // Required methods
    fn next_target(&self) -> Option<NodeId>;
    fn set_next_target(&mut self, target: Option<NodeId>);
}

The NextTarget trait provides methods for managing a single target for a railway object.

Required Methods

fn next_target(&self) -> Option<NodeId>

Returns the next target of the railway object, if any.

fn set_next_target(&mut self, target: Option<NodeId>)

Sets the next target of the railway object.