openrailwaymap_exporter

Module simulation

This module contains types related to railway simulations.

A railway simulation consists of a railway graph representing the infrastructure and a list of movable railway objects, such as trains, within the simulation. The module provides a Simulation struct to manage the state of the simulation.

Re-exports

pub use environment::SimulationEnvironment;

Modules

agents

This module contains types and traits related to decision agents.

commands

This module provides a set of commands that can be used to manipulate a Simulation instance. Each command implements the SimulationCommand trait, which provides the execute method to apply the command on a given simulation instance.

environment

module with environment traits

events

This module contains types related to simulation events in railway simulations.

metrics

This module contains types related to handling and processing metrics during railway simulations.

Structs

Simulation

A Simulation struct holding a railway graph and a list of moveable railway objects.

SimulationExecutor

A struct to execute a simulation with a specific frame rate and runtime.

Traits

SimulationObject

A trait that defines an object within the simulation that can move along a railway.