openrailwaymap_exporter::simulation

Module commands

This module provides a set of commands that can be used to manipulate a Simulation instance. Each command implements the SimulationCommand trait, which provides the execute method to apply the command on a given simulation instance.

Structs

PauseCommand

A command that toggles the pause state of a simulation.

SetSpeedupCommand

A command that sets the speedup factor of the simulation.

Enums

MetricsCommand

An enumeration of subcommands for working with metrics in the simulation.

ObjectCommand

A command for querying object information in a simulation.

Traits

SimulationCommand

A trait for commands that manipulate a simulation.