Enum openrailwaymap_exporter::simulation::commands::ObjectCommand

pub enum ObjectCommand {
    List,
    Show(ShowObject),
}

A command for querying object information in a simulation.

This command has two subcommands:

list - Lists the IDs of all objects in the simulation.

show - Shows detailed information about an object with a specified ID.

Example usage:

// List all objects
let list_cmd = ObjectCommand::List;
let list_output = list_cmd.execute(&mut simulation);

// Show information about an object with ID 1
let show_cmd = ObjectCommand::Show(ShowObject { object_id: 1 });
let show_output = show_cmd.execute(&mut simulation);

Variants

List

Lists the IDs of all objects in the simulation.

Show(ShowObject)

Shows detailed information about an object with a specified ID.