Struct openrailwaymap_exporter::simulation::environment::SimulationEnvironment

pub struct SimulationEnvironment {
    pub graph: RailwayGraph,
    pub objects: HashMap<i64, Box<dyn SimulationObject>>,
}

Represents the environment for the decision agent, containing the RailwayGraph.

Fields
graph: RailwayGraph

A reference to the RailwayGraph in the environment.

objects: HashMap<i64, Box<dyn SimulationObject>>

A collection of simulation objects, keyed by their unique identifiers.