openrailwaymap_exporter::simulation::environment

Trait ObservableEnvironmentRef

pub trait ObservableEnvironmentRef {
    // Required method
    fn as_observable_env(&self) -> &(dyn ObservableEnvironment + 'static);
}

A trait representing a reference to an observable environment.

Implementations of this trait should provide a method to obtain a reference to an ObservableEnvironment object.

Required Methods

fn as_observable_env(&self) -> &(dyn ObservableEnvironment + 'static)

Returns a reference to an ObservableEnvironment object.