openrailwaymap_exporter::simulation

Module events

This module contains types related to simulation events in railway simulations.

Simulation events are used to represent specific occurrences during the execution of a simulation, such as a train changing its action or reaching a target.

The module provides a SimulationEvent trait that defines the interface for all simulation events. The RailMovableEvent and TargetReachedEvent structs are example implementations of this trait, representing a change in a RailMovableAction and a target being reached, respectively.

Structs

RailMovableEvent

A RailMovableEvent represents a change in a RailMovableAction during a simulation.

TargetReachedEvent

A TargetReachedEvent represents a target being reached during a simulation.

Traits

SimulationEvent

The SimulationEvent trait defines the interface for all simulation events.