Struct openrailwaymap_exporter::RailwayGraph

pub struct RailwayGraph {
    pub graph: Graph<RailwayNode, RailwayEdge, Undirected>,
    pub node_indices: HashMap<i64, NodeIndex>,
}

RailwayGraph represents a graph structure for railway networks.

The graph consists of nodes representing railway stations and junctions, and edges representing the railway segments between the nodes. It also stores the node indices as a HashMap for easy retrieval.

Fields
graph: Graph<RailwayNode, RailwayEdge, Undirected>

The internal graph used to represent the railway network.

The graph consists of RailwayNode instances as nodes and RailwayEdge instances as edges. It is an undirected graph.

node_indices: HashMap<i64, NodeIndex>

A HashMap that maps node IDs to their corresponding indices in the graph.

This HashMap allows for quick and easy retrieval of node indices based on their IDs.

Implementations

impl RailwayGraph

pub fn bounding_box(&self) -> (Coordinate, Coordinate)

Calculate the bounding box of the graph.

The bounding box is represented as a tuple containing the minimum and maximum latitude and longitude values of the nodes in the graph.

Returns

A tuple containing two Coordinate structs representing the minimum and maximum coordinates of the bounding box of the graph.

pub fn from_railway_elements(elements: &[RailwayElement]) -> Self

Create a RailwayGraph from a vector of RailwayElements.

The function processes the input elements to create a graph with nodes and edges.

Arguments

elements - A vector of RailwayElements from which the graph will be created.

Returns

A RailwayGraph created from the input RailwayElements.

Example

use openrailwaymap_exporter::{ElementType, RailwayElement};
use openrailwaymap_exporter::Coordinate;
use openrailwaymap_exporter::RailwayGraph;
use std::collections::HashMap;

let elements = vec![
    RailwayElement {
        id: 1,
        element_type: ElementType::Node,
        lat: Some(50.1109),
        lon: Some(8.6821),
        tags: Some(HashMap::new()),
        nodes: None,
        geometry: None,
    },
    RailwayElement {
        id: 2,
        element_type: ElementType::Way,
        lat: None,
        lon: None,
        tags: Some(HashMap::new()),
        nodes: Some(vec![1, 3]),
        geometry: None,
    },
];

let railway_graph = RailwayGraph::from_railway_elements(&elements);
println!("Created railway graph with {} nodes", railway_graph.graph.node_count());