Trait openrailwaymap_exporter::simulation::agents::DecisionAgent

pub trait DecisionAgent: Send {
    type A;

    fn next_action(&self, delta_time: Option<Duration>) -> Self::A;
    fn observe(&mut self, environment: &SimulationEnvironment);
}

A trait that represents a decision agent responsible for choosing the best action based on the current state of the simulation.

Required Associated Types

type A

The associated action type for this decision agent.

Required Methods

fn next_action(&self, delta_time: Option<Duration>) -> Self::A

Returns the best action based on the current state of the simulation.

Returns

Self::A - The action chosen by the decision agent.

fn observe(&mut self, environment: &SimulationEnvironment)

Observes the current environment and updates the agent’s internal state.

Arguments

environment - The current environment.