Trait openrailwaymap_exporter::simulation::events::SimulationEvent

pub trait SimulationEvent {
    // Required method
    fn as_any(&self) -> &dyn Any;
}

The SimulationEvent trait defines the interface for all simulation events.

Required Methods

fn as_any(&self) -> &dyn Any

Returns a reference to the event as a trait object implementing Any. This method is useful for downcasting the event to a concrete type.